Fix Relay generation schema script to delete temporary file if there's an exception

+++ bin/generateSchemaAndRunRelayCompiler.ts
@@ -4,7 +4,15 @@ import { execSync } from 'child_process'
 
 const existingSchemaFile = 'src/main/resources/relay/schema.graphql'
 
-const combinedSchema = getGraphqlSchema()
-fs.writeFileSync(existingSchemaFile, combinedSchema)
-execSync("relay-compiler")
-fs.unlinkSync(existingSchemaFile)
+try {
+  const combinedSchema = getGraphqlSchema()
+  fs.writeFileSync(existingSchemaFile, combinedSchema)
+  execSync("relay-compiler")
+} catch (e) {
+  console.error("Relay compilation failed with error: " + e)
+}
+finally {
+  if (fs.existsSync(existingSchemaFile)) {
+    fs.unlinkSync(existingSchemaFile)
+  }
+}
